moving edit timetable function to separate file
