+++ entities/views/__init__.py
@@ -1,3 +1,4 @@
 from .security import *
 from .views import *
 from .generate_page import *
+from .edit_timetable import *
